Save highlights to file
Add highlight mode autopage (selected with -m) which is like page, with the addition
that highlights are saved when the file is closed (or program exits), to a file
named as the pdf file concatenated with '.pymp'. This file is read when a pdf file is
openned.
File format is json, with Gdk.RGBA object replaced by a dictionary with single key 'rgba'
and value a quartet (r,g,b,a).