Add option to ignore small movements when scribbling
The option min_distance in presenter section in the config file sets the (square of) the shortest segment that may be added to a scribble.
This gives a tradeoff between performance on one hand and exactness and smoothness of the curve on the other.
It makes sense especially for writing pads which usually have a very high resolution, resulting in thousands of segments even for short scribbles.
The units are page width/height, and number to experiment should usually be between 0.000001 and 0.001.